requirements.txt
# --- 웹 서버 / REST API ---
flask>=2.3.0
requests>=2.31.0

# --- ML ---
scikit-learn>=1.4.0
joblib>=1.3.0
numpy>=1.26.0
pandas>=2.2.0

# --- LLM / RAG ---
openai>=1.10.0
faiss-cpu>=1.7.4          # (Windows에서 문제 있으면 chromadb 대체 가능인데 굳이 싶음)
chromadb>=0.5.0            # FAISS 안 쓰면 대신 사용
tqdm>=4.66.0

# --- 환경설정 / 경로 / 유틸 ---
python-dotenv>=1.0.1
pathlib>=1.0.1

# --- 테스트 ---
pytest>=8.0.0

# --- 선택적 (시각화 / 개발 편의) ---
gunicorn>=21.2.0           # 서버 배포용 (optional)
pyflakes>=3.2.0            # 코드 검사 (optional)
